refactor: move CenterButton and RouteButton to a new MapButtons component
